append

+++ UI/CitrusTouchUI.h
@@ -2,7 +2,7 @@
 //  CitrusTouchUI.h
 //  CitrusTouch2017
 //
-//  Created by kouhei.takemoto on 2017/03/28.
+//  Created by take64 on 2017/03/28.
 //  Copyright © 2017年 citrus.live. All rights reserved.
 //
 
@@ -10,11 +10,11 @@
 #define CitrusTouchUI_h
 
 
-__attribute__((unused)) static CGFloat CT8(int hex)
+__attribute__((unused)) static CGFloat CT8(CGFloat hex)
 {
     return hex * 8;
 }
-__attribute__((unused)) static NSString * CT8Str(int hex)
+__attribute__((unused)) static NSString * CT8Str(CGFloat hex)
 {
     return [@(CT8(hex)) stringValue];
 }
@@ -29,10 +29,21 @@ __attribute__((unused)) static NSString * CTStr(CGFloat floatValue)
 
 #import "CTControl.h"
 #import "CTLabel.h"
+#import "CTButton.h"
 
 #import "CTView.h"
 
+// overlay
+#import "CTOverlayProgressIndicator.h"
+
 // table cell
 #import "CTTableCell.h"
+#import "CTTableCellButton.h"
+#import "CTTableCellLabel.h"
+#import "CTTableCellTextField.h"
+#import "CTTableCellTextFieldInnerTextField.h"
+
+// table header footer
+#import "CTTableHeaderFooterView.h"
 
 #endif /* CitrusTouchUI_h */
